moving ValidateSample function and tests to ingester package. It's only referenced by the ingester struct and with the addition of flags for the label name/value lengths, it makes more sense for it to live in the ingester pkg
