+++ pkg/ingester/ingester.go
@@ -329,7 +329,7 @@ func (i *Ingester) append(ctx context.Context, sample *model.Sample) error {
 		return httpgrpc.Errorf(http.StatusBadRequest, "sample with timestamp %v is older than the maximum accepted age", sample.Timestamp)
 	}
 
-	if err := util.ValidateSample(sample); err != nil {
+	if err := ValidateSample(sample); err != nil {
 		level.Error(util.WithContext(ctx, util.Logger)).Log("msg", "error validating sample", "err", err)
 		return nil
 	}
